feat: #133 支持Github系列-添加友好提示
