+++ README.md
@@ -67,8 +67,6 @@ https://mp.terwer.space/post/start-to-get-started-zi0eyk.html
 |:-------------------------------------------------------------------------------------:|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------:|:----:|
 |  ![](https://static-rs-terwer.oss-cn-beijing.aliyuncs.com/icons/browser/edge20.svg)   | [v0.0.3](https://microsoftedge.microsoft.com/addons/detail/%E6%80%9D%E6%BA%90%E7%AC%94%E8%AE%B0%E5%8F%91%E5%B8%83%E8%BE%85%E5%8A%A9%E5%B7%A5%E5%85%B7/aejmkigifflimhjlhjkdckclhabbilee) | 已上架  |
 | ![](https://static-rs-terwer.oss-cn-beijing.aliyuncs.com/icons/browser/chrome20.svg)  |     [v0.0.3](https://chrome.google.com/webstore/detail/%E6%80%9D%E6%BA%90%E7%AC%94%E8%AE%B0%E5%8F%91%E5%B8%83%E8%BE%85%E5%8A%A9%E5%B7%A5%E5%85%B7/gemlnnppcphbiimfjnobfgdkohjmgifm)     | 已上架  |
-| ![](https://static-rs-terwer.oss-cn-beijing.aliyuncs.com/icons/browser/firefox20.png) |                                                                                           ---                                                                                           | 审核中  |
-|  ![](https://static-rs-terwer.oss-cn-beijing.aliyuncs.com/icons/browser/360-20.png)   |                                                                                           ---                                                                                           | 审核中  |
 |                                          网页版                                          |                                                                 [v0.0.3](https://publish.terwer.space/blog/index.html)                                                                  | 可使用  |
 |                                         离线zip                                         |                                  [v0.0.3](https://github.com/terwer/src-sy-post-publisher/releases/download/v0.0.3/sy-post-publisher-chrome-0.0.3.zip)                                  | 可下载  |
 |                                          挂件版                                          |                                                                  [v0.0.3](https://github.com/terwer/sy-post-publisher)                                                                  | 可使用  |
@@ -76,8 +74,6 @@ https://mp.terwer.space/post/start-to-get-started-zi0eyk.html
 小贴士：
 
 1. 推荐直接从 `思源笔记集市`、 `Chrome商店` 或者 `Edge商店` 等官方商店下载使用。
-
-   由于技术限制，`Firefox插件` 目前不支持 `Minifist V3` 以及 `background messsage` 事件这个关键特性，需要额外使用请求代理，参照 [#143](https://github.com/terwer/src-sy-post-publisher/issues/143) 。这样的话，性能会差一点，但是功能完全可用。代理地址为：https://publish.terwer.space/api/middleware
 
 2. 如果您想 `临时尝鲜新版无需等待版本发布` 可以从网页版直接使用，网页版访问地址：https://publish.terwer.space/blog/index.html
